refactor(tuning): drop redundant casts in _holdout_maes
Same two redundancies e10038f removed from censorship.py: astype(float)
followed by to_numpy() is one to_numpy(dtype=float), and the float() around
each MAE is undone by the np.asarray(dtype=float) on the way out.

+++ src/retail_forecasting/forecasting/imputation_tuning.py
@@ -55,8 +55,8 @@ def _holdout_maes(holdouts: list[Holdout], params: dict[str, int | float]) -> np
     maes = []
     for censored, eval_idx, true_demand in holdouts:
         imputed = LatentDemandImputer(strategy="supervised", lgbm_params=params).impute(censored)
-        pred = imputed.loc[eval_idx, "latent_demand_est"].astype(float).to_numpy()
-        maes.append(float(np.mean(np.abs(pred - true_demand))))
+        pred = imputed.loc[eval_idx, "latent_demand_est"].to_numpy(dtype=float)
+        maes.append(np.mean(np.abs(pred - true_demand)))
     return np.asarray(maes, dtype=float)
